refactor: extract plugin configuration logic into a dedicated file
